requirements.txt
streamlit
spacy
pytrends
textblob
openai
fpdf
pandas
scikit-learn
matplotlib
python-dotenv